Fix slow print jobs on Dymo
This is a hack I've used before, and it works quite well. I did play
around with all kinds of CUPS settings, such as
`usb-unidir-default=true` and `usb-no-reattach-default=true`, but
I cannot find a "proper" fix for this.

Killing the stalled process after a job is complete works every time.